Updated differential burst fractions in energetics.py to be 0.0 below Emin, not 1.0
JXP had changed these from 0.0 to 1.0 below Emin, maybe to be consistent with the cumulative burst fractions (?). In the case of cumulative burst fractions, 1 is correct because the total number of bursts observed with energy greater than Eth for Eth<Emin is always one. Whereas, for the differential case, I think 0.0 is correct as the differential functions are calculating the change, as a function of energy, in the cumulative number of bursts observed above some Eth, for Eth < Emin, which is zero, because for all Eth < Emin the cumulative number is 1.